+++ src/adapters/android_backlight.h
@@ -31,12 +31,12 @@ public:
     AndroidBacklight();
     ~AndroidBacklight();
 
-    void set_brightness(float) override;
-    float get_brightness() override;
+    void set_brightness(double) override;
+    double get_brightness() override;
 
 private:
     light_device_t* light_dev;
-    float brightness;
+    double brightness;
 };
 
 }
